+++ test/case-assemblematrix.sh
@@ -1,0 +1,3 @@
+
+./xsvutils ./test/data/assemblematrix.tsv assemblematrix  > ./var/test/assemblematrix-1.tsv
+