+++ CHANGELOG.md
@@ -9,6 +9,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - `-` means stdin
 - add input file name validation
 - remove -v4 option of `rmnoname` subcommand
+- add `assemblematrix` subcommand
 
 ## version 0.4 (2018/07/10)
 - support -v1,-v2,-v3,-v4